+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+### Fixes
+
+* *Simple condition:* an error occurs when the field is of type camelcase or pascalcase  (e.g., `input.streetNumber`, `input.StreetNumber`).
+
+
 ## [0.8.0] - July, 2024
 
 ### Features